tests/wpt/tests/html/semantics/embedded-content/media-elements/track/track-element/track-webvtt-newlines.html

<!DOCTYPE html>
<title>A cue with no newline at eof is parsed properly</title>
<script src="track-helpers.js"></script>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<video>
    <track src="resources/no-newline-at-eof.vtt" default>
    <script>
        async_test(function(t) {
            var track = document.querySelector("track");

            track.onload = t.step_func_done(function() {
                var expected = [
                    {
                        id : "1",
                        startTime : 0,
                        endTime : 30.5,
                        text : "Bear is Coming!!!!!"
                    }
                ];

                assert_cues_equal(track.track.cues, expected);
            });
        });
    </script>
</video>
